feat: breaking change footers
Adds a prompt, and the functionality, for adding a breaking change footer describing the breaking changes in detail, as described in the Conventional Commits specification.
I decided against using a dash (-) between "BREAKING" and "CHANGE" as I fear that not every tool supports it, but it may be configurable in the future.

Closes #86