feat: git-like -C argument, integration tests (#103)
Added an argument almost equivalent to Git's "-C", mainly for the
purpose of making integration tests isolated. This feature depends on
cocogitto/cocogitto#428, and therefore v6.2.0, as the commit will
otherwise still be made in the current working directory.
This commit reduces code test coverage as many skips were removed.